requirements.txt
pandas
numpy
matplotlib
nbinteract
scipy
nibabel
nilearn
jupyter-book
pyyaml